1. Moved to a superclass model for common functionality 2. Gemini support seems to be (at least mostly) working... and we have metadata now too!
